# neo4j_vkg/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Purple Fabric - Enterprise Knowledge Graph Platform API
  description: |
    REST API for the Purple Fabric Enterprise Knowledge Graph Platform.
    Provides document ingestion, ontology management, graph querying, entity resolution,
    multi-tenant workspace management, and AI-powered extraction services.

    ## Authentication
    All endpoints (except `/api/auth/login`, `/api/auth/refresh`, `/api/auth/logout`, and `/api/health`)
    require a Bearer JWT token in the `Authorization` header.

    Access tokens expire in 15 minutes. Use the refresh token endpoint to obtain new tokens.

    ## CSRF Protection
    All state-changing requests (POST, PUT, DELETE) must include the `X-Requested-With` header.

    ## Role-Based Access Control (RBAC)
    Four roles with hierarchical permissions:
    - **viewer** — Read-only access (dashboards, queries, chat)
    - **member** — Upload docs, run extractions, contribute content
    - **manager** — Manage ontologies, clear data, organize workspace
    - **admin** — Full access including user management and system settings

    ## Multi-Tenant Architecture
    Resources are scoped to tenants and workspaces. Pass `tenant_id` and `workspace_id`
    via query parameters or the `X-Tenant-ID` / `X-Workspace-ID` headers.
  version: 1.0.0
  contact:
    name: Purple Fabric Team
servers:
  - url: http://localhost:5002
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenant_id
      in: query
      schema: { type: string }
      description: Tenant ID for multi-tenant scoping
    WorkspaceId:
      name: workspace_id
      in: query
      schema: { type: string }
      description: Workspace ID for workspace scoping
...
security:
  - BearerAuth: []
paths:
...
  /api/ontology/{id}:
    get:
      tags: [Ontology]
      summary: Get a specific ontology by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Ontology details }
        '404': { description: Ontology not found }
  /api/ontology/templates:
    get:
      tags: [Ontology]
      summary: Get available ontology templates
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200': { description: Templates and custom ontologies }
  /api/ontology/templates/{id}:
    get:
      tags: [Ontology]
      summary: Get a specific template
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Template details }
        '404': { description: Template not found }
...
  /api/folders/{id}/ontology:
    get:
      tags: [Folders]
      summary: Get folder's linked ontology
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Linked ontology }
    post:
      tags: [Folders]
      summary: Link ontology to folder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Ontology linked }

  # ============================================================
  # TENANTS
  # ============================================================
...